fix: resolve codacy analysis
