+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/MovieDataSourcePostTest.kt
@@ -4,8 +4,7 @@ import com.waffiq.bazz_movies.core.network.data.remote.models.FavoritePostModel
 import com.waffiq.bazz_movies.core.network.data.remote.models.RatePostModel
 import com.waffiq.bazz_movies.core.network.data.remote.models.WatchlistPostModel
 import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.post.PostResponse
-import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.OMDbApiService
-import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.TMDBApiService
+import com.waffiq.bazz_movies.core.network.testutils.BaseMovieDataSourceTest
 import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.postResponseSuccessDump
 import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.ratePostResponseSuccessDump
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
@@ -16,21 +15,14 @@ import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testIOExceptionR
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testSocketTimeoutExceptionResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testSuccessResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testUnknownHostExceptionResponse
-import com.waffiq.bazz_movies.core.test.MainDispatcherRule
-import io.mockk.MockKAnnotations
-import io.mockk.clearMocks
-import io.mockk.impl.annotations.MockK
 import junit.framework.TestCase.assertEquals
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.runTest
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.ResponseBody.Companion.toResponseBody
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import retrofit2.Response
 
-class MovieDataSourcePostTest {
+class MovieDataSourcePostTest : BaseMovieDataSourceTest() {
 
   private val userId = 123123
   private val sessionId = "session_id"
@@ -54,32 +46,6 @@ class MovieDataSourcePostTest {
     """{"status_code": 422, "status_message": "Invalid parameters: Your request parameters are incorrect."}"""
       .toResponseBody("application/json".toMediaTypeOrNull())
   )
-
-  @MockK
-  private lateinit var tmdbApiService: TMDBApiService
-
-  @MockK
-  private lateinit var omDbApiService: OMDbApiService
-
-  @MockK
-  private lateinit var testDispatcher: Dispatchers
-
-  @get:Rule
-  val mainDispatcherRule = MainDispatcherRule()
-
-  private lateinit var movieDataSource: MovieDataSource
-
-  @Before
-  fun setup() {
-    // Initialize MockK annotations and relax mocking behavior
-    // `relaxed = true` allows MockK to automatically provide default behavior for any unmocked method calls.
-    MockKAnnotations.init(this, relaxed = true)
-
-    // clear any previous mocks to ensure tests are isolated
-    clearMocks(tmdbApiService, omDbApiService)
-
-    movieDataSource = MovieDataSource(tmdbApiService, omDbApiService, testDispatcher.IO)
-  }
 
   @Test
   fun postFavorite_ReturnExpectedResponse() = runTest {
