+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/UserDataSourceLoginTest.kt
@@ -1,11 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.datasource
 
-import com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.post.PostResponse
-import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.CountryIPApiService
-import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.TMDBApiService
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.accountDetailsResponse
+import com.waffiq.bazz_movies.core.network.testutils.BaseUserDataSourceTest
 import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.authenticationResponseDump
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.countryIPResponseDump
 import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.createSessionResponseDump
 import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.postResponseDump
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
@@ -16,49 +12,13 @@ import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testIOExceptionR
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testSocketTimeoutExceptionResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testSuccessResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testUnknownHostExceptionResponse
-import com.waffiq.bazz_movies.core.test.MainDispatcherRule
-import io.mockk.MockKAnnotations
-import io.mockk.clearMocks
-import io.mockk.impl.annotations.MockK
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertTrue
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.test.runTest
-import okhttp3.MediaType.Companion.toMediaTypeOrNull
-import okhttp3.ResponseBody.Companion.toResponseBody
-import org.junit.Before
-import org.junit.Rule
 import org.junit.Test
 import retrofit2.Response
 
-class UserDataSourceTest {
-
-  private val apiInvalidFormatErrorResponse: Response<PostResponse> = Response.error(
-    405,
-    """{"status_code": 503, "status_message": "Invalid format: This service doesn't exist in that format."}"""
-      .toResponseBody("application/json".toMediaTypeOrNull())
-  )
-
-  @MockK
-  private lateinit var tmdbApiService: TMDBApiService
-
-  @MockK
-  private lateinit var countryIPApiService: CountryIPApiService
-
-  @MockK
-  private lateinit var testDispatcher: Dispatchers
-
-  @get:Rule
-  val mainDispatcherRule = MainDispatcherRule()
-
-  private lateinit var userDataSource: UserDataSource
-
-  @Before
-  fun setup() {
-    MockKAnnotations.init(this, relaxed = true)
-    clearMocks(tmdbApiService, countryIPApiService)
-    userDataSource = UserDataSource(tmdbApiService, countryIPApiService, testDispatcher.IO)
-  }
+class UserDataSourceLoginTest : BaseUserDataSourceTest() {
 
   @Test
   fun createToken_ReturnExpectedResponse() = runTest {
@@ -280,154 +240,6 @@ class UserDataSourceTest {
     )
   }
   // endregion createSessionLogin EDGE CASE
-
-  @Test
-  fun getUserDetail_ReturnExpectedResponse() = runTest {
-    testSuccessResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      mockApiResponse = Response.success(accountDetailsResponse),
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-      expectedData = accountDetailsResponse,
-    ) { data ->
-      assertEquals(543798538, data.id)
-      assertEquals("en", data.iso6391)
-      assertEquals("USERNAME", data.username)
-      assertTrue(data.includeAdult == false)
-    }
-  }
-
-  @Test
-  fun getUserDetail_ReturnExpectedStatusMessageResponse() = runTest {
-    testErrorResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      errorResponse = apiInvalidFormatErrorResponse,
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-      expectedErrorMessage = "Invalid format: This service doesn't exist in that format."
-    )
-  }
-
-  // region getUserDetail EDGE CASE
-  @Test
-  fun getUserDetail_ReturnErrorWhenAPIRespondsWith404() = runTest {
-    testError404Response(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-
-  @Test
-  fun getUserDetail_ReturnErrorWhenNetworkErrorOccurs() = runTest {
-    testUnknownHostExceptionResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-
-  @Test
-  fun getUserDetail_ReturnErrorWhenTimeoutOccurs() = runTest {
-    testSocketTimeoutExceptionResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-
-  @Test
-  fun getUserDetail_ReturnErrorWhenHttpExceptionOccurs() = runTest {
-    testHttpExceptionResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-
-  @Test
-  fun getUserDetail_ReturnErrorWhenIOExceptionOccurs() = runTest {
-    testIOExceptionResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-
-  @Test
-  fun getUserDetail_ReturnErrorWhenExceptionOccurs() = runTest {
-    testGeneralExceptionResponse(
-      apiEndpoint = { tmdbApiService.getAccountDetails("session_id") },
-      dataSourceEndpointCall = { userDataSource.getUserDetail("session_id") },
-    )
-  }
-  // endregion getUserDetail EDGE CASE
-
-  @Test
-  fun getCountryCode_ReturnExpectedResponse() = runTest {
-    testSuccessResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      mockApiResponse = Response.success(countryIPResponseDump),
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-      expectedData = countryIPResponseDump,
-    ) { data ->
-      assertEquals("ID", data.country)
-      assertEquals("103.187.242.255", data.ip)
-    }
-  }
-
-  @Test
-  fun getCountryCode_ReturnExpectedStatusMessageResponse() = runTest {
-    testErrorResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      errorResponse = apiInvalidFormatErrorResponse,
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-      expectedErrorMessage = "Invalid format: This service doesn't exist in that format."
-    )
-  }
-
-  // region getCountryCode EDGE CASE
-  @Test
-  fun getCountryCode_ReturnErrorWhenAPIRespondsWith404() = runTest {
-    testError404Response(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-
-  @Test
-  fun getCountryCode_ReturnErrorWhenNetworkErrorOccurs() = runTest {
-    testUnknownHostExceptionResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-
-  @Test
-  fun getCountryCode_ReturnErrorWhenTimeoutOccurs() = runTest {
-    testSocketTimeoutExceptionResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-
-  @Test
-  fun getCountryCode_ReturnErrorWhenHttpExceptionOccurs() = runTest {
-    testHttpExceptionResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-
-  @Test
-  fun getCountryCode_ReturnErrorWhenIOExceptionOccurs() = runTest {
-    testIOExceptionResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-
-  @Test
-  fun getCountryCode_ReturnErrorWhenExceptionOccurs() = runTest {
-    testGeneralExceptionResponse(
-      apiEndpoint = { countryIPApiService.getIP() },
-      dataSourceEndpointCall = { userDataSource.getCountryCode() },
-    )
-  }
-  // endregion getCountryCode EDGE CASE
 
   @Test
   fun login_ReturnExpectedResponse() = runTest {
